personalization_pilot/requirements.txt
streamlit
glob2
gspread
google-auth
oauth2client